+++ docs/changes.rst
@@ -12,6 +12,8 @@ Not yet released.
 
 **Bug fixes**
 
+* Fixed translations caching in :ref:`machine-translation-setup`.
+
 **Compatibility**
 
 **Upgrading**